moved weather functions to their own separate file
